checkin

+++ cloud-app/src/main/java/com/njit/aryeh/TextRecognition2.java
@@ -33,18 +33,20 @@ public class TextRecognition2 {
 				for (Message message : messages) {
 					String messageBody = message.body();
 					System.out.println("message: " + messageBody);
-					DeleteMessageRequest deleteMessageRequest = DeleteMessageRequest.builder()
-                            .queueUrl(queueUrl)
-                            .receiptHandle(message.receiptHandle())
-                            .build();
-                    sqsClient.deleteMessage(deleteMessageRequest);
-					if(messageBody.equals("-1")) {
+
+					if (false) {
+						DeleteMessageRequest deleteMessageRequest = DeleteMessageRequest.builder().queueUrl(queueUrl)
+								.receiptHandle(message.receiptHandle()).build();
+						sqsClient.deleteMessage(deleteMessageRequest);
+					}
+
+					if (messageBody.equals("-1")) {
 						System.out.println("-1 received ... exiting");
 						exitLoop = true;
 						break;
 					}
 				}
-				if(exitLoop) {
+				if (exitLoop) {
 					System.out.println("exiting loop");
 					break;
 				}
